+++ regions/shapes/line.py
@@ -84,7 +84,7 @@ class LinePixelRegion(PixelRegion):
     def to_sky(self, wcs):
         start = pixel_to_skycoord(self.start.x, self.start.y, wcs)
         end = pixel_to_skycoord(self.end.x, self.end.y, wcs)
-        return LineSkyRegion(start, end)
+        return LineSkyRegion(start, end, meta=self.meta, visual=self.visual)
 
     @property
     def bounding_box(self):
@@ -198,4 +198,4 @@ class LineSkyRegion(SkyRegion):
         start = PixCoord(start_x, start_y)
         end_x, end_y = skycoord_to_pixel(self.end, wcs=wcs)
         end = PixCoord(end_x, end_y)
-        return LinePixelRegion(start, end)
+        return LinePixelRegion(start, end, meta=self.meta, visual=self.visual)