+++ regions/shapes/tests/test_text.py
@@ -9,7 +9,7 @@ import astropy.units as u
 from astropy.utils.data import get_pkg_data_filename
 from astropy.wcs import WCS
 
-from ...core import PixCoord
+from ...core import PixCoord, RegionMeta, RegionVisual
 from ...tests.helpers import make_simple_wcs
 from ..text import TextPixelRegion, TextSkyRegion
 from .test_common import BaseTestPixelRegion, BaseTestSkyRegion
@@ -23,8 +23,10 @@ def wcs_fixture():
 
 
 class TestTextPixelRegion(BaseTestPixelRegion):
-
-    reg = TextPixelRegion(PixCoord(3, 4), "Sample Text")
+    meta = RegionMeta({'text': 'test'})
+    visual = RegionVisual({'color': 'blue'})
+    reg = TextPixelRegion(PixCoord(3, 4), "Sample Text", meta=meta,
+                          visual=visual)
     sample_box = [-2, 8, -1, 9]
     inside = []
     outside = [(3.1, 4.2), (5, 4)]
@@ -38,14 +40,16 @@ class TestTextPixelRegion(BaseTestPixelRegion):
         reg = self.reg.copy()
         assert reg.center.xy == (3, 4)
         assert reg.text == "Sample Text"
-        assert reg.visual == {}
-        assert reg.meta == {}
+        assert reg.meta == self.meta
+        assert reg.visual == self.visual
 
     def test_pix_sky_roundtrip(self):
         wcs = make_simple_wcs(SkyCoord(2 * u.deg, 3 * u.deg), 0.1 * u.deg, 20)
         reg_new = self.reg.to_sky(wcs).to_pixel(wcs)
         assert_allclose(reg_new.center.x, self.reg.center.x)
         assert_allclose(reg_new.center.y, self.reg.center.y)
+        assert reg_new.meta == self.reg.meta
+        assert reg_new.visual == self.reg.visual
 
     def test_rotate(self):
         reg = self.reg.rotate(PixCoord(2, 3), 90 * u.deg)
@@ -53,8 +57,10 @@ class TestTextPixelRegion(BaseTestPixelRegion):
 
 
 class TestTextSkyRegion(BaseTestSkyRegion):
-
-    reg = TextSkyRegion(SkyCoord(3, 4, unit='deg'), "Sample Text")
+    meta = RegionMeta({'text': 'test'})
+    visual = RegionVisual({'color': 'blue'})
+    reg = TextSkyRegion(SkyCoord(3, 4, unit='deg'), "Sample Text",
+                        meta=meta, visual=visual)
     expected_repr = ('<TextSkyRegion(center=<SkyCoord (ICRS): (ra, dec) in '
                      'deg\n    (3., 4.)>, text=Sample Text)>')
     expected_str = ('Region: TextSkyRegion\ncenter: <SkyCoord (ICRS): '
@@ -64,8 +70,8 @@ class TestTextSkyRegion(BaseTestSkyRegion):
         reg = self.reg.copy()
         assert_allclose(reg.center.ra.deg, 3)
         assert reg.text == "Sample Text"
-        assert reg.visual == {}
-        assert reg.meta == {}
+        assert reg.meta == self.meta
+        assert reg.visual == self.visual
 
     def test_contains(self, wcs):
         position = SkyCoord([1, 2] * u.deg, [3, 4] * u.deg)